archive/agents/ddaas-web-agent/requirements.txt
azure.ai.projects
azure.identity